fix: changed test case location
